# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Biblioteca",
        "description": "Sistema de Gerenciamento de Biblioteca",
        "contact": {
            "name": "Equipe de Suporte",
            "email": "suporte@biblioteca.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Servidor de API da Biblioteca"
        }
    ],
    "paths": {
...
        "/api/v1/loans": {
            "get": {
                "tags": [
                    "Empréstimos"
                ],
                "summary": "Lista todos os empréstimos",
                "operationId": "11067bb9b862fd7effd5a9394c6aa29e",
                "responses": {
                    "200": {
                        "description": "Lista de empréstimos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Loan"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Empréstimos"
                ],
                "summary": "Criar novo empréstimo",
                "operationId": "67746018747d878b72f133d469781040",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoanCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empréstimo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Loan"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Livro não disponível para empréstimo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/v1/loans/{id}": {
            "get": {
                "tags": [
                    "Empréstimos"
                ],
                "summary": "Obter detalhes de um empréstimo específico",
                "operationId": "93e76696ff0b06ec950dd75e35c89c1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do empréstimo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Loan"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empréstimo não encontrado"
                    }
                }
            }
        },
        "/api/v1/loans/{id}/return": {
            "get": {
                "tags": [
                    "Empréstimos"
                ],
                "summary": "Devolver livro",
                "operationId": "e26749263f5e4c0918cd8ab2acc646bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livro devolvido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empréstimo não encontrado"
                    }
                }
            }
        },
        "/api/v1/loans/{id}/mark-as-delayed": {
            "get": {
                "tags": [
                    "Empréstimos"
                ],
                "summary": "Marcar empréstimo como atrasado",
                "operationId": "f8311c4d96b4bfd9277f6faac2dd1e0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empréstimo marcado como atrasado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empréstimo não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Loan": {
                "title": "Empréstimo",
                "description": "Modelo de empréstimo de livro",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "book_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "delayed",
                            "returned"
                        ],
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "LoanCreateRequest": {
                "title": "Requisição de Criação de Empréstimo",
                "required": [
                    "user_id",
                    "book_id"
                ],
                "properties": {
                    "user_id": {
                        "description": "ID do usuário que está fazendo o empréstimo",
                        "type": "integer",
                        "example": 1
                    },
                    "book_id": {
                        "description": "ID do livro a ser emprestado",
                        "type": "integer",
                        "example": 1
                    },
                    "due_date": {
                        "description": "Data de devolução (opcional)",
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}